refactored io module to use xarray accessor instead of subclassing
